studenten aanmaken en deleten

+++ Project/src/timetable/create/CreateController.java
@@ -6,10 +6,7 @@ import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.geometry.Pos;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableCell;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.control.cell.TextFieldTableCell;
 import javafx.scene.layout.AnchorPane;
@@ -20,6 +17,7 @@ import timetable.db.DataAccessException;
 import timetable.objects.Item;
 
 import java.io.IOException;
+import java.util.Optional;
 
 public class CreateController {
     private Stage stage;
@@ -35,7 +33,8 @@ public class CreateController {
     @FXML
     private TableView<Item> table;
     @FXML
-    private Button student, teacher, loc, lecture;
+    private Button student, teacher, loc, lecture, create;
+    private Boolean canClose=true;
 
     public void setStageAndSetupListeners(Stage stage, Controller mainController) {
         //krijgen van de stage
@@ -124,27 +123,15 @@ public class CreateController {
     }
 
 
-    public void saveStudent() {
-        if (!name.getText().isEmpty()) {
-            try (DataAccessContext dac = mainController.getModel().getDataAccessProvider().getDataAccessContext()) {
-                dac.getStudentsDAO().createStudent(name.getText());
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
+    public void create() {
+        try (DataAccessContext dac = mainController.getModel().getDataAccessProvider().getDataAccessContext()) {
+            table.getItems().add(dac.getStudentsDAO().createStudent("student"));
+            mainController.getModel().changeItems(mainController.getModel().getStandardSchedule());
+        } catch (Exception e) {
+            e.printStackTrace();
         }
-        mainController.getModel().fireInvalidationEvent();
     }
 
-    public void saveTeacher() {
-        if (!name.getText().isEmpty()) {
-            try (DataAccessContext dac = mainController.getModel().getDataAccessProvider().getDataAccessContext()) {
-                dac.getStudentsDAO().createStudent(name.getText());
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-        mainController.getModel().fireInvalidationEvent();
-    }
 
     public void menu() {
         try {
@@ -158,14 +145,31 @@ public class CreateController {
     }
 
     public void close() {
-        stage.close();
+        if(canClose){
+            stage.close();
+        }
     }
 
     private void delete(Item item) {
-        try (DataAccessContext dac = mainController.getModel().getDataAccessProvider().getDataAccessContext()) {
-            dac.getStudentsDAO().deleteStudent(item);
-        } catch (Exception e) {
-            e.printStackTrace();
+        canClose = false;
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation Dialog");
+        alert.setHeaderText("Are you sure");
+        alert.setContentText("If a student is used in a lecture,\n then that lecture gets deleted too ");
+
+        Optional<ButtonType> result = alert.showAndWait();
+        if (result.isPresent() && result.get() == ButtonType.OK){
+            // ... user chose OK
+
+            try (DataAccessContext dac = mainController.getModel().getDataAccessProvider().getDataAccessContext()) {
+                dac.getStudentsDAO().deleteStudent(item);
+                mainController.getModel().changeItems(mainController.getModel().getStandardSchedule());
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        } else {
+            // ... user chose CANCEL or closed the dialog
         }
+        canClose = true;
     }
 }
